Manual

FunnyTN

FunnyTN.MPO — Type.

MPO{T} <: TensorTrain

Matrix Product Operator.

We use the following convention to number legs: 2 | 1–A–4 | 3

llink -> 1 ulink -> 2 dlink -> 3 rlink -> 4

source

FunnyTN.MPOTensor — Type.

MPO Tensor

source

FunnyTN.MPSTensor — Type.

MPS Tensor

source

FunnyTN.vec2mps — Method.

Parse a normal state into a Matrix produdct state.

state: The target state, 1D array. nflavor: The dimension of a single site, integer. l: The division point of left and right canonical scanning, integer between 0 and number of site. method: The method to extract orthonormal matrices. * :qr -> get A,B matrices by the method of QR decomposition, faster, rank revealing in a non-straight-forward way. * :svd -> get A,B matrices by the method of SVD decomposition, slow, rank revealing. tol: The tolerence of singular value, float.

Return an <MPS> instance.

source

FunnyTN.bond — Function.

bond(tt::AbstractTN, l::Int) -> Bond

l-th bound of tensor train.

source

FunnyTN.log2i — Function.

log2i(x::Integer) -> Integer

Return log2(x), with integer input only.

source

FunnyTN.tensors — Function.

tensors(tt::AbstractTN) -> Tensor

Get the list tensors.

source